refactor(admin/semesters): SemesterRegisterDrawer 를 react-hook-form + components 분리로 전환
- useState + handcrafted handlers → useForm + Controller + register/setValue
- HeroUI v3 Select 의 selectedKey/onSelectionChange 가 wrapper 타입에 노출되지
  않아 무시되던 이슈를 Controller + ListBox.Item onClick 패턴으로 해결
- 단일 ~780줄 파일을 components 폴더의 7개 파일로 분리:
  SemesterRegisterDrawer / BasicInfoSection / ProcessSection /
  CurriculumSection / ApplicationFormSection / DeleteCohortDialog / shared
- 기존 components/Sections.tsx (하드코딩 dead code) 삭제
