+++ apps/admin/src/pages/semesters/components/ProcessSection.tsx
@@ -1,0 +1,235 @@
+import {
+  Calendar,
+  DateField,
+  DatePicker,
+  DateRangePicker,
+  RangeCalendar,
+} from "@heroui/react"
+import { Controller } from "react-hook-form"
+import type { Control, UseFormSetValue } from "react-hook-form"
+
+import { GridBox } from "@/shared/ui/GridBox"
+
+import type { SemesterRegisterForm } from "../types"
+
+import { FormField, SectionTitle } from "./shared"
+
+interface Props {
+  control: Control<SemesterRegisterForm>
+  setValue: UseFormSetValue<SemesterRegisterForm>
+}
+
+export function ProcessSection({ control, setValue }: Props) {
+  return (
+    <section className="space-y-4">
+      <SectionTitle>프로세스 일정</SectionTitle>
+      <GridBox className="grid-cols-2 gap-5">
+        <FormField label="서류 접수">
+          <DateRangePicker
+            className="w-full"
+            value={null}
+            onChange={(value) => {
+              setValue(
+                "process.documentAcceptStartDate",
+                value?.start.toString() ?? "",
+                { shouldDirty: true },
+              )
+              setValue(
+                "process.documentAcceptEndDate",
+                value?.end.toString() ?? "",
+                { shouldDirty: true },
+              )
+            }}
+          >
+            <DateField.Group fullWidth>
+              <DateField.Input slot="start">
+                {(segment) => <DateField.Segment segment={segment} />}
+              </DateField.Input>
+              <DateRangePicker.RangeSeparator />
+              <DateField.Input slot="end">
+                {(segment) => <DateField.Segment segment={segment} />}
+              </DateField.Input>
+              <DateField.Suffix>
+                <DateRangePicker.Trigger>
+                  <DateRangePicker.TriggerIndicator />
+                </DateRangePicker.Trigger>
+              </DateField.Suffix>
+            </DateField.Group>
+            <DateRangePicker.Popover>
+              <RangeCalendar aria-label="서류 접수 기간">
+                <RangeCalendar.Header>
+                  <RangeCalendar.YearPickerTrigger>
+                    <RangeCalendar.YearPickerTriggerHeading />
+                    <RangeCalendar.YearPickerTriggerIndicator />
+                  </RangeCalendar.YearPickerTrigger>
+                  <RangeCalendar.NavButton slot="previous" />
+                  <RangeCalendar.NavButton slot="next" />
+                </RangeCalendar.Header>
+                <RangeCalendar.Grid>
+                  <RangeCalendar.GridHeader>
+                    {(day) => (
+                      <RangeCalendar.HeaderCell>{day}</RangeCalendar.HeaderCell>
+                    )}
+                  </RangeCalendar.GridHeader>
+                  <RangeCalendar.GridBody>
+                    {(date) => <RangeCalendar.Cell date={date} />}
+                  </RangeCalendar.GridBody>
+                </RangeCalendar.Grid>
+              </RangeCalendar>
+            </DateRangePicker.Popover>
+          </DateRangePicker>
+        </FormField>
+
+        <FormField label="서류 발표">
+          <Controller
+            control={control}
+            name="process.documentResultDate"
+            render={({ field }) => (
+              <DatePicker
+                className="w-full"
+                value={null}
+                onChange={(date) => field.onChange(date?.toString() ?? "")}
+              >
+                <DateField.Group fullWidth>
+                  <DateField.Input>
+                    {(segment) => <DateField.Segment segment={segment} />}
+                  </DateField.Input>
+                  <DateField.Suffix>
+                    <DatePicker.Trigger>
+                      <DatePicker.TriggerIndicator />
+                    </DatePicker.Trigger>
+                  </DateField.Suffix>
+                </DateField.Group>
+                <DatePicker.Popover>
+                  <Calendar aria-label="서류 발표">
+                    <Calendar.Header>
+                      <Calendar.YearPickerTrigger>
+                        <Calendar.YearPickerTriggerHeading />
+                        <Calendar.YearPickerTriggerIndicator />
+                      </Calendar.YearPickerTrigger>
+                      <Calendar.NavButton slot="previous" />
+                      <Calendar.NavButton slot="next" />
+                    </Calendar.Header>
+                    <Calendar.Grid>
+                      <Calendar.GridHeader>
+                        {(day) => (
+                          <Calendar.HeaderCell>{day}</Calendar.HeaderCell>
+                        )}
+                      </Calendar.GridHeader>
+                      <Calendar.GridBody>
+                        {(date) => <Calendar.Cell date={date} />}
+                      </Calendar.GridBody>
+                    </Calendar.Grid>
+                  </Calendar>
+                </DatePicker.Popover>
+              </DatePicker>
+            )}
+          />
+        </FormField>
+
+        <FormField label="인터뷰 날짜">
+          <DateRangePicker
+            className="w-full"
+            value={null}
+            onChange={(value) => {
+              setValue(
+                "process.interviewStartDate",
+                value?.start.toString() ?? "",
+                { shouldDirty: true },
+              )
+              setValue(
+                "process.interviewEndDate",
+                value?.end.toString() ?? "",
+                { shouldDirty: true },
+              )
+            }}
+          >
+            <DateField.Group fullWidth>
+              <DateField.Input slot="start">
+                {(segment) => <DateField.Segment segment={segment} />}
+              </DateField.Input>
+              <DateRangePicker.RangeSeparator />
+              <DateField.Input slot="end">
+                {(segment) => <DateField.Segment segment={segment} />}
+              </DateField.Input>
+              <DateField.Suffix>
+                <DateRangePicker.Trigger>
+                  <DateRangePicker.TriggerIndicator />
+                </DateRangePicker.Trigger>
+              </DateField.Suffix>
+            </DateField.Group>
+            <DateRangePicker.Popover>
+              <RangeCalendar aria-label="인터뷰 기간">
+                <RangeCalendar.Header>
+                  <RangeCalendar.YearPickerTrigger>
+                    <RangeCalendar.YearPickerTriggerHeading />
+                    <RangeCalendar.YearPickerTriggerIndicator />
+                  </RangeCalendar.YearPickerTrigger>
+                  <RangeCalendar.NavButton slot="previous" />
+                  <RangeCalendar.NavButton slot="next" />
+                </RangeCalendar.Header>
+                <RangeCalendar.Grid>
+                  <RangeCalendar.GridHeader>
+                    {(day) => (
+                      <RangeCalendar.HeaderCell>{day}</RangeCalendar.HeaderCell>
+                    )}
+                  </RangeCalendar.GridHeader>
+                  <RangeCalendar.GridBody>
+                    {(date) => <RangeCalendar.Cell date={date} />}
+                  </RangeCalendar.GridBody>
+                </RangeCalendar.Grid>
+              </RangeCalendar>
+            </DateRangePicker.Popover>
+          </DateRangePicker>
+        </FormField>
+
+        <FormField label="최종 발표">
+          <Controller
+            control={control}
+            name="process.finalResultDate"
+            render={({ field }) => (
+              <DatePicker
+                className="w-full"
+                value={null}
+                onChange={(date) => field.onChange(date?.toString() ?? "")}
+              >
+                <DateField.Group fullWidth>
+                  <DateField.Input>
+                    {(segment) => <DateField.Segment segment={segment} />}
+                  </DateField.Input>
+                  <DateField.Suffix>
+                    <DatePicker.Trigger>
+                      <DatePicker.TriggerIndicator />
+                    </DatePicker.Trigger>
+                  </DateField.Suffix>
+                </DateField.Group>
+                <DatePicker.Popover>
+                  <Calendar aria-label="최종 발표">
+                    <Calendar.Header>
+                      <Calendar.YearPickerTrigger>
+                        <Calendar.YearPickerTriggerHeading />
+                        <Calendar.YearPickerTriggerIndicator />
+                      </Calendar.YearPickerTrigger>
+                      <Calendar.NavButton slot="previous" />
+                      <Calendar.NavButton slot="next" />
+                    </Calendar.Header>
+                    <Calendar.Grid>
+                      <Calendar.GridHeader>
+                        {(day) => (
+                          <Calendar.HeaderCell>{day}</Calendar.HeaderCell>
+                        )}
+                      </Calendar.GridHeader>
+                      <Calendar.GridBody>
+                        {(date) => <Calendar.Cell date={date} />}
+                      </Calendar.GridBody>
+                    </Calendar.Grid>
+                  </Calendar>
+                </DatePicker.Popover>
+              </DatePicker>
+            )}
+          />
+        </FormField>
+      </GridBox>
+    </section>
+  )
+}
