+++ apps/admin/src/pages/semesters/components/index.ts
@@ -1,0 +1,2 @@
+export { SemesterRegisterDrawer } from "./SemesterRegisterDrawer"
+export type { DrawerMode } from "./SemesterRegisterDrawer"
